# src/main/java/com/sapient/domain/ProductDto.java
package com.sapient.domain;

public class ProductDto {
    private Integer id;
    private String name;
    private String type;
    private String price;

    public ProductDto() {
    }

    public ProductDto(String name, String type, String price) {
        this.name = name;
        this.type = type;
        this.price = price;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }

    public String getPrice() {
        return price;
    }

    public void setPrice(String price) {
        this.price = price;
    }

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }
}